+++ library/baseline/csr.hpp
@@ -110,7 +110,7 @@ private:
 protected:
     // Helper, translate the logical into real vertices IDs. Materialization step at the end of a graphalytics algorithm
     template <typename T>
-    std::vector<std::pair<uint64_t, T>> translate(T* values, int N);
+    std::vector<std::pair<uint64_t, T>> translate(T* values, uint64_t N);
 
     // Helper, save the content of the vector to the given output file
     template <typename T, bool negative_scores = true>
